+++ jax/experimental/pallas/__init__.py
@@ -56,7 +56,7 @@ from jax._src.pallas.pallas_call import pallas_call_p as pallas_call_p
 from jax._src.pallas.primitives import debug_print as debug_print
 from jax._src.pallas.primitives import delay as delay
 from jax._src.pallas.primitives import DeviceIdType as DeviceIdType
-from jax._src.pallas.primitives import dot as dot
+from jax._src.pallas.primitives import dot as _deprecated_dot
 from jax._src.pallas.primitives import get_global as get_global
 from jax._src.pallas.primitives import multiple_of as multiple_of
 from jax._src.pallas.primitives import num_programs as num_programs
@@ -81,10 +81,23 @@ ANY = MemorySpace.ANY
 HOST = MemorySpace.HOST
 
 _deprecations = {
+    # Added June 4, 2026
+    "dot": (
+        (
+            "jax.experimental.pallas.dot was moved to"
+            " jax.experimental.pallas.triton. Accessing it via"
+            " jax.experimental.pallas is deprecated. You can use jax.numpy.dot,"
+            " jax.numpy.einsum or the @ operator instead in a TPU or MGPU"
+            " kernel."
+        ),
+        _deprecated_dot,
+    ),
     # Added May 15, 2026
     "debug_checks_enabled": (
-        "jax.experimental.pallas.debug_checks_enabled is deprecated, "
-        "use pl.enable_debug_checks.value instead.",
+        (
+            "jax.experimental.pallas.debug_checks_enabled is deprecated, "
+            "use pl.enable_debug_checks.value instead."
+        ),
         _deprecated_debug_checks_enabled,
     ),
 }
@@ -92,6 +105,7 @@ _deprecations = {
 import typing
 if typing.TYPE_CHECKING:
   debug_checks_enabled = _deprecated_debug_checks_enabled
+  dot = _deprecated_dot
 else:
   from jax._src.deprecations import deprecation_getattr as _deprecation_getattr
   __getattr__ = _deprecation_getattr(__name__, _deprecations)